# generated-openapi-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trieve API",
    "description": "Trieve OpenAPI Specification. This document describes all of the operations available through the Trieve API.",
    "contact": {
      "name": "Trieve Team",
      "url": "https://trieve.ai",
      "email": "developers@trieve.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/devflowinc/actix-web-template/blob/main/LICENSE.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "create_task",
        "parameters": [
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to create a new task",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON body representing the task that was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Service error relating to the request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "operationId": "get_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object representing the requested task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "operationId": "update_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to update the task",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Object representing the updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Service error relating to the request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "operationId": "delete_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the task was successfully deleted"
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/tasks/{task_id}/{resource_type}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "operationId": "list_task_resource",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PrefixedUuid"
            }
          },
          {
            "name": "resource_type",
            "in": "path",
            "description": "The resource type to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TaskResType"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The number of records to skip",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of objects of resource_type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResourceListWithPagination"
                }
              }
            }
          },
          "400": {
            "description": "Service error relating to the request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/tasks/{task_id}/{resource_type}/{resource_id}": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "create_task_resource",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PrefixedUuid"
            }
          },
          {
            "name": "resource_type",
            "in": "path",
            "description": "The resource type to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TaskResType"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "The resource id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Object representing the created relationship",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResource"
                }
              }
            }
          },
          "400": {
            "description": "Service error relating to the request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "operationId": "delete_task_resource",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PrefixedUuid"
            }
          },
          {
            "name": "resource_type",
            "in": "path",
            "description": "The resource type to use for the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TaskResType"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "The resource id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Organization",
            "in": "header",
            "description": "The organization id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the task resource was successfully deleted"
          },
          "400": {
            "description": "Service error relating to the request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTaskReqPayload": {
        "type": "object",
        "properties": {
          "contact_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PrefixedUuid"
              }
            ],
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-01T00:00:00",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ErrorRespPayload": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "Bad Request"
        }
      },
...
      "Link": {
        "type": "object",
        "required": [
          "id",
          "link",
          "org_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "link": {
            "type": "string"
          },
          "org_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2021-01-01T00:00:00",
          "id": "link-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "link": "https://trieve.ai",
          "org_id": "org-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
...
      "PrefixedUuid": {
        "type": "string"
      },
      "Task": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "contact_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PrefixedUuid"
              }
            ],
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "org_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "contact_id": "contact-c7c7c7c7-c7c7-c7c7-c7c7-c7c7c7c7c7c7",
          "created_at": "2021-01-01T00:00:00",
          "deadline": "2021-01-01T00:00:00",
          "description": "Task description...",
          "id": "task-c7c7c7c7-c7c7-c7c7-c7c7-c7c7c7c7c7c7",
          "org_id": "org-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
      "TaskDeal": {
        "type": "object",
        "required": [
          "id",
          "task_id",
          "deal_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deal_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "task_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2021-01-01T00:00:00",
          "deal_id": "deal-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "id": "taskdeal-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "task_id": "task-c7c7c7c7-c7c7-c7c7-c7c7-c7c7c7c7c7c7",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
      "TaskLink": {
        "type": "object",
        "required": [
          "id",
          "task_id",
          "link_id",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "link_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "task_id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2021-01-01T00:00:00",
          "id": "tasklink-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "link_id": "link-b8b8b8b8-b8b8-b8b8-b8b8-b8b8b8b8b8b8",
          "task_id": "task-c7c7c7c7-c7c7-c7c7-c7c7-c7c7c7c7c7c7",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
      "TaskResType": {
        "type": "string",
        "enum": [
          "Link",
          "Deal",
          "User"
        ]
      },
      "TaskResource": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Link"
            ],
            "properties": {
              "Link": {
                "$ref": "#/components/schemas/TaskLink"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Deal"
            ],
            "properties": {
              "Deal": {
                "$ref": "#/components/schemas/TaskDeal"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "User"
            ],
            "properties": {
              "User": {
                "$ref": "#/components/schemas/TaskUser"
              }
            }
          }
        ]
      },
      "TaskResourceList": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Link"
            ],
            "properties": {
              "Link": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Link"
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Deal"
            ],
            "properties": {
              "Deal": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Deal"
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "User"
            ],
            "properties": {
              "User": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        ]
      },
      "TaskResourceListWithPagination": {
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskResourceList"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateTaskReqPayload": {
        "type": "object",
        "properties": {
          "contact_id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PrefixedUuid"
              }
            ],
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "example": "2021-01-01T00:00:00",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
...
}